pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mm-video"
version = "0.2.0"
description = "A lightweight and flexible training framework with highly strcutured configuration system based on Hydra"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    # configuration
    "hydra-core", "omegaconf",
    "fvcore", # will be removed in the future
    # basic
    "tensorboard",
    "tqdm",
    "joblib>=1.3.0",
    # log
    "tabulate", "colorlog",
    # deep learning
    "torch>=2.0.0", "torchvision",
    "einops",
    "numpy", "pandas",
    "pillow", "opencv-python",
]

[tool.setuptools.packages.find]
exclude = ["test*", "model_zoo*", "dataset*", "configs*", "scripts*", "tools*"]

[tool.whell]
exclude = ["test*", "model_zoo*", "dataset*", "configs*", "scripts*", "tools*"]